requirements.txt
fastapi~=0.85.0
requests-toolbelt~=0.9
rowantree.auth.service~=0.12
starlette~=0.20
